requirements.txt
numpy
torch
transformers
datasets
botorch 
gpytorch
scipy
accelerate
pymoo
openai
python-dotenv
# xformers # memory_efficient_attention to accelerate training